+++ src/components/Home/Home.css
@@ -1,0 +1,29 @@
+.expenses {
+    padding: 1rem;
+    background-color: rgb(31, 31, 31);
+    width: 100%;
+    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
+    display:flex;
+}
+.ml-1 {
+    margin-left: 1vw
+}
+.hub {
+    display: block;
+    font-family: sans-serif;
+    font-weight: bold;
+    font-size: 4vw;
+    top: 50%;
+    left: 50%;
+    margin-right: 1vw;
+    }
+.hub span:nth-child(1) {
+    color: #FFFFFF;
+    }
+.hub span:nth-child(2) {
+    background: #FF9900;
+    color: #000000;
+    border-radius: 1vw;
+    padding: 0 1vw 1vw 1vw;
+    display: inline-block;
+    }
